pyproject.toml
[tool.poetry]
name = "interview_cleaned"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pyspark = "3.5.1"
pandas = "2.2.2"
pyjanitor = "0.27.0"
jupyterlab = "4.1.6"
ipython = "8.23.0"
scikit-learn = "1.4.2"
apache-airflow = "2.8.4"
pyarrow = "16.0.0"
boto3 = "1.34.88"
sqlalchemy = "1.4.52"
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
